Try to find remaining issue with numpy 1.x + win combination
A problem occurs in filter_test.py::test_logic that is suggestive (but not conclusive) of the int32 vs int64 issue

+++ pynbody/snapshot/subsnap.py
@@ -137,7 +137,7 @@ class SubSnapBase(SimSnap):
 
     def get_index_list(self, relative_to, of_particles=None):
         if of_particles is None:
-            of_particles = np.arange(len(self))
+            of_particles = np.arange(len(self), dtype=np.int64)
 
         if relative_to is self:
             return of_particles
